+++ src/Screens/ScreenConnexion/index.js
@@ -1,10 +1,11 @@
 import React from 'react'
+import styles from '../../Composants/styles/styles';
+import ButtonComponent from '../../Composants/Bouton/buttonComponent';
 import { useState } from 'react'
-import { TextInput, StyleSheet, View, Text, Alert } from 'react-native'
+import { TextInput, View, Text, Alert } from 'react-native'
 import { SafeAreaView } from 'react-native-safe-area-context';
 import { login } from '../../../api';
-import BoutonApp from '../../Composants/Bouton'
-import Footer from '../../Composants/Footer';
+
 import Header from '../../Composants/Header'
 
 export default function ConnexionScreen({ navigation }) {
@@ -39,7 +40,7 @@ export default function ConnexionScreen({ navigation }) {
         <SafeAreaView style={styles.container}>
             <Header nav={navigation} />
             <Text style={styles.title}>Identifiez-vous</Text>
-            <View style={styles.formContainer}>
+            <View style={styles.contenerCenter}>
                 <TextInput
                     style={styles.input}
                     onChangeText={onChangeEmail}
@@ -54,39 +55,15 @@ export default function ConnexionScreen({ navigation }) {
                     placeholder='Password'
                     secureTextEntry={true}
                 />
-                <BoutonApp text="Connexion"
-                    onPress={handleSubmit}
-                />
+          
+                <ButtonComponent 
+                contButon={styles.contenerCenter}
+                button={styles.butonStyle}
+                txtButton={styles.textButon}
+                text={"Connexion"}
+                onPress={handleSubmit}
+            />
             </View>
-            <Footer />
         </SafeAreaView>
     )
 }
-const styles = StyleSheet.create({
-    container: {
-        flex: 1,
-        alignItems: "center",
-        backgroundColor: "#0077B6",
-    },
-    formContainer: {
-        alignItems: 'center',
-        width: "80%",
-    },
-    title: {
-        color: '#FFFFFF',
-        fontSize: 40,
-        margin: 30,
-        textAlign: 'center',
-        fontFamily: 'Iceland_400Regular'
-    },
-    input: {
-        backgroundColor: "#F0F0F0",
-        borderColor: 'black',
-        borderWidth: 1,
-        borderRadius: 3,
-        width: '80 %',
-        height: "12%",
-        margin: '5%',
-        paddingLeft: 15,
-    },
-})
