+++ src/Screens/ScreenIndexArticles/readArticle.js
@@ -1,4 +1,3 @@
-
 import React, { useState, useEffect, Fragment } from 'react'
 import { Text, StyleSheet, ScrollView, View } from 'react-native';
 import { SafeAreaView } from 'react-native-safe-area-context';
@@ -8,7 +7,8 @@ import Header from '../../Composants/Header';
 import BoutonApp from '../../Composants/Bouton';
 import ModalAddComment from '../../Composants/Modals/ModalAddComment';
 import { isLogged } from '../../../lib';
-
+import styles from '../../Composants/styles/styles';
+import ButtonComponent from '../../Composants/Bouton/buttonComponent';
 
 export default function ReadArticle({ route, navigation }) {
     const articleId = route.params
@@ -53,7 +53,7 @@ export default function ReadArticle({ route, navigation }) {
             <ScrollView style={{ marginBottom: "10%" }}>
                 {article && article.comments ? (
                     <View>
-                        <Text style={styles.title}>{article.title}</Text>
+                        <Text style={styles.titleH2}>{article.title}</Text>
                         <View style={styles.Author}>
                             <Text style={styles.date}>Le {brassageDate(article.createdAt)}</Text>
                             <Text style={styles.date}>Par : {article.userId.lastname + ' ' + article.userId.firstname}</Text>
@@ -62,7 +62,18 @@ export default function ReadArticle({ route, navigation }) {
                             <Text style={styles.txt}>{article.content}</Text>
                </View>
                             
-                        {log ? (<Fragment><Text style={styles.date}>Commentaires</Text><BoutonApp style={styles.btn} text="Add"  onPress={openAdd} /></Fragment>) : (<Text style={styles.carreful}>Connectez-vous pour ajouter un commentaire !</Text>)}
+                        {log ? (<Fragment>
+                        <View style={styles.OneLine}>
+                        <Text style={styles.date}>Commentaires</Text>
+                        <ButtonComponent 
+                            contButon={styles.date}
+                            button={styles.butonStyleLitte}
+                            txtButton={styles.textButon}
+                            text={"Add"}
+                            onPress={openAdd}
+                        />
+                        </View>
+                        </Fragment>) : (<Text style={styles.carreful}>Connectez-vous pour ajouter un commentaire !</Text>)}
                 
                         {isOpenAdd && <ModalAddComment close={openAdd} onPress={openAdd} id={articleId.articleId} />}
                        
@@ -84,57 +95,3 @@ export default function ReadArticle({ route, navigation }) {
         </SafeAreaView>
     )
 }
-const styles = StyleSheet.create({
-    container: {
-        flex: 1,
-        backgroundColor: "#0077B6",
-    },
-    title: {
-        color: "#FFFFFF",
-        fontSize: 30,
-        marginTop: 20,
-        textAlign: "center",
-        fontFamily: 'Iceland_400Regular'
-    },
-    Author: {
-        flexDirection: "row",
-        justifyContent: "space-around",
-    },
-    date: {
-        color: "#FFFFFF",
-        fontSize: 16,
-        marginTop: "5%",
-    },
-    articleContainer: {
-        marginVertical: "7%",
-        height: "auto",
-        paddingVertical: "10%",
-        paddingHorizontal: "3%"
-    },
-    txt: {
-        color: "#FFFFFF",
-        fontSize: 20,
-        textAlign: "justify"
-    },
-    commentsContainer: {
-        marginTop: "5%"
-    },
-    OneComments: {
-        backgroundColor: "#90E0EF",
-        marginHorizontal: "3%",
-        marginBottom: "5%",
-        padding: "2%",
-        borderRadius: 5
-    },
-    carreful: {
-        fontSize: 15,
-        alignItems: "center",
-        alignContent: "center",
-        textAlign: "center",
-        padding: 10,
-        backgroundColor: "#48CAE4",
-        marginVertical: "7%",
-        height: "auto",
-
-    }
-})
