+++ src/Screens/ScreenProfilPage/index.js
@@ -1,11 +1,15 @@
 import React, { useState, useEffect } from 'react'
 import { Text, View, StyleSheet, ScrollView } from 'react-native'
 import { SafeAreaView } from 'react-native-safe-area-context';
-
+import styles from '../../Composants/styles/styles';
+import ButtonComponent from '../../Composants/Bouton/buttonComponent';
 import BoutonApp from '../../Composants/Bouton'
 import { getMe } from '../../../api';
 import Header from '../../Composants/Header'
 import Footer from '../../Composants/Footer';
+
+import { login } from '../../../api';
+// Faire la logique/
 
 export default function ProfilScreen({ navigation }) {
 
@@ -25,14 +29,21 @@ export default function ProfilScreen({ navigation }) {
             <Header nav={navigation} />
             <ScrollView style={{ marginBottom: "15%", width: "90%" }}>
                 <Text style={styles.title}>Profile</Text>
-                <View style={styles.inputContainer}>
-                    <Text style={styles.txt}>Email: {user.email}</Text>
-                    <Text style={styles.txt}>FirstName: {user.firstname}</Text>
-                    <Text style={styles.txt}>LastName: {user.lastname}</Text>
-                    <Text style={styles.txt}>Rôle: </Text>
+                <View style={styles.contenerLeft}>
+                    <Text style={styles.txt}>Email: {user.email}
+                    </Text>
+                    <Text style={styles.txt}>
+                        FirstName: {user.firstname}
+                    </Text>
+                    <Text style={styles.txt}>
+                        LastName: {user.lastname}
+                    </Text>
+                    <Text style={styles.txt}>
+                        Rôle: 
+                    </Text>
                 </View>
                 <Text style={styles.titleH2}>Modifier votre profile</Text>
-                <View style={styles.inputContainer}>
+                <View style={styles.contenerLeft}>
                     <Text style={styles.txt}>Email:</Text>
                     <Text style={styles.txt}>Password:</Text>
                     <Text style={styles.txt}>FirstName:</Text>
@@ -45,34 +56,3 @@ export default function ProfilScreen({ navigation }) {
         </SafeAreaView>
     )
 }
-const styles = StyleSheet.create({
-    container: {
-        flex: 1,
-        alignItems: "center",
-        backgroundColor: "#0077B6",
-    },
-    inputContainer: {
-        marginHorizontal: "5%",
-        marginBottom: "3%"
-    },
-    title: {
-        color: '#FFFFFF',
-        fontSize: 45,
-        marginTop: 30,
-        marginBottom: 10,
-        textAlign: 'center',
-        fontFamily: 'Iceland_400Regular'
-    },
-    titleH2: {
-        color: '#FFFFFF',
-        fontSize: 35,
-        margin: 15,
-        textAlign: 'center',
-        fontFamily: 'Iceland_400Regular'
-    },
-    txt: {
-        color: "#fff",
-        fontSize: 20,
-        marginTop: 20
-    }
-})
